testing/web-platform/mozilla/tests/indic-detection/elango.html

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
ÙNYëYÖš (Mars) s¡Veh|•T†‡¥ E·[ J£ ÚLÖ· Bh•. C‰ s¡V]ìXì£‹‰ SÖÁLÖY‰ ÚLÖ[ÖL E·[‰.Ceh|•T†‡¥ –Lop½V ÚLÖ[Ö] “RÄeh A|†RRÖL CWQëPÖY‰ p½V ÚLÖ[ÖL ÙNYëYÖš C£ef\‰. ÚU]Öyz]Ÿ CeÚLÖºeh ÚTÖŸeLP°¸Á ÙTVÛWo syz·[]Ÿ. CRÁ ÚU¼TWTë‘¥ LÖQTëT|• C£•“ BeÛN| CeÚLÖÛ[o ÙN‹Sì\UÖLe LÖy|f\‰.[12] CR]ÖÚXÚV CR¼ho ÙNYëYÖš GÁ\ ÙTVŸ H¼TyP‰. J£ “«NÖŸ ÚLÖ[Ö] CRÁ ÚU¼TWTë“ N‹‡W]ì¥ E·[‰ÚTÖ¥ fQëQe hZìLÛ[•, “«›¥ E·[‰ ÚTÖÁ\ G¡UÛXL·, T·[†RÖehL·, TÖÛXY]jL·, T]ì™zV ‰£YTë Th‡LÛ[• ÙLÖQëP‰. ÙNYëYÖ›Á rZ¼peLÖX˜•, T£Y UÖ¼\jLº• “«eh E·[ÛRTë ÚTÖÁ\ÛYÚV. s¡V UQëPX†‰· –L EVWUÖ] JXì•Tr UÛX•, –LTëÙT¡V ÙNjh†‰Të T·[†RÖehLº· JÁ\Ö] U¡]Ÿ T·[†RÖeh• ÙNYëYÖ›ÚXÚV E·[].
<script>
test(function() {
  assert_equals(document.characterSet, "windows-1252");
},"Should fall back to windows-1252");
</script>
The text content above is a converted extract from the start of the Wikipedia article <a href="https://ta.wikipedia.org/w/index.php?title=%E0%AE%9A%E0%AF%86%E0%AE%B5%E0%AF%8D%E0%AE%B5%E0%AE%BE%E0%AE%AF%E0%AF%8D_(%E0%AE%95%E0%AF%8B%E0%AE%B3%E0%AF%8D)&oldid=3129711">&#2970;&#3014;&#2997;&#3021;&#2997;&#3006;&#2991;&#3021; (&#2965;&#3019;&#2995;&#3021;)</a> and
is licensed under the <a href="https://creativecommons.org/licenses/by-sa/3.0/legalcode">Creative Commons Attribution-ShareAlike 3.0 Unported</a> license.
